Create README.md
Create a photograph of five people sitting in a bombed private house near a campfire. Here is the description of the people:

The first person is missing their left leg and has a prosthetic in its place. They are wearing the military uniform of Ukraine.

The second person should have developed muscles, dressed in black clothing, wearing a bandana, and armed with an AK-74U. They should resemble a looter.

The third person is dressed in technical military gear from Russia and is sitting while fixing a drone in their hands.

The other two should be depicted as civilians.